add logging to databroker and move from module code
